Added a browse page-type to the server controller map.
This allows us to understand what a browse page-type is for the services
page.

`parse` in StagecraftApiClient needs to have a truthy value. This lets
the result be happily parsed, and then the services or homepage
controllers can carry on processing the data. Feels a little icky, but
works for now.