Switch everything to use flatten=true parameter
We were down to the last module as part of the switch over to flat
group_by data, so I migrated the addition of the flatten=true query
parameter from each individual module/collection into the data source.
This should make pulling it a little easier and ensures that all
requests will have this parameter provided (as everything uses
data_source).

This maintains the ability of a particular module to override this query
parameter as it as the json from stagecraft and any custom query
parameters get merged over the top.