Switch on flatten for everything
I've hidden this behind a flag that is passed through to DataSource from
any instance of Collection. This has been done so that we don't have to
immediately move all of the test fixtures to representing flatten data
rather than the nested data. This will be done but decoupled from
enabling flatten for all.