Use dashboard generator to update dashboards
The dashboard generator was originally written to take dashboards
from a Google Spreadsheet and save them as JSON files.

We now need to pull in changes from the Google Spreadsheet and
update the data stored in Stagecraft.

This commit retrieves the database ID for a given dashboard slug
and uses that to ensure the dashboard is overwritten in the
database rather than being duplicated.