User sat graphs should be more configurable
In order to migrate the user_satisfaction_graph module over to using
transformed data we need more control over the keys that the module uses
to pull data out of records.

This also rewrites the gathering of a breakdown to use the axes rather
than a hard coded list of labels.

The migrated property is to skip the application of query parameters and
parsing that is required for dealing with the untransformed data.