User totalAttr to pull out total responses
totalAttr looked like to should be used to define which key would derive
the total response from. This was not actually being used in the
function and the totalRatings key exists no where. I have now defaulted
it to the correct field name and used it where it is appropriate.