Fix ```grunt jasmine_node:server``` command
If you run this command at moment it will fail as there's no global logger defined.

It works if you run the main jasmine_node as the helpers for ```legacy``` add it to the global.

This adds it to the global for server so you can now just run the server tests only and be happy.