Fix line breaks in legend labels on hover
The labels in the legend for graphs like https://www.gov.uk/performance/register-to-vote/registrations-by-age-group cause a slightly nasty overlap because a line break appears on hover.
Setting the right value at 0 allows the absolutely positioned labels to fill the full width of the parent container, and so prevents the line breaks appearing when the font-face is changed to bold.