Fix formatOptions problems.
When format is specified as `format: 'currency'` shorthand the property is not used.

This fix puts it in line with how formatting is done elsewhere on spotlight.

If format is specified like so:

`format: 'percent'`

Spotlight mixes it into the format defaults.

If it's:

```
format: {
  type: 'percent'
}
```

It will also mix that into the defaults.

Added tests around this as there were none.