Homepage - check if showcase service is found before adding
There are 4 predefined 'showcase' services to display at the top of the home page - 'prison-visits', 'book-practical-driving-test', 'carers-allowance', 'renew-patent'
There wasn't a proper check that each of these was present in the list of dashboards from the API, before adding to a list for rendering. This has now been corrected and error logging improved.