Fixup issue with the client serving 404s for all errors
When stagecraft returns something to the client that isn't a 404 (502 for example) then the client will always query the local filesystem for a dashboard.

This is bad as the resulting error message will be a 404.

This change removes the fetchFallback (local filesystem check for a dashboard) and just errors with the response from stagecraft to the end user.