Specific caching directives for error HTML
If an attempt to render a dashboard fails, we should not cache the
response for the normal period. Instead, allow a small window to allow
the system to recover if it’s under load.

Tests needed a bit of poking to understand. Spy usage means that simple
data structures aren’t being used in some places; eg PageConfig.