Parse config file as Mustache
This commit parses the config file as Mustache and replaces instances
of `govukAppDomain` with the environment variable GOVUK_APP_DOMAIN.

This is useful for our preview and staging config files because instead
of hardcoding `alphagov.co.uk` we can easily move to a new hostname.

The syntax on lines 4 and 6 is a little bit confusing. What it's doing
is reassigning the Mustache tokens to be `<% %>` so that on line 5
we can include Mustache which isn't parsed (literal Mustache tags).

These tags are parsed later in the application, in `data_source.js:26`.