Handle no data for total cost and no of transactions
Previously we would always initialize the aggregrates for the services
page with 0s, however if no data was found for an service page aggregate
kpi then this would show up as 0 on the page itself in the case of
number of transactions and total cost because of the way these avoid the weighted_average logic.

This changes the logic so that if no kpi is yet defined we first check
there is actually volume information before adding it to aggregates.
This has the problem that even if there is a value it will not be
counted in the weighted average because we cannot say what proportion of
the average it represents but it seems better then displaying data when
none is actually provided. We leave a special case for total_cost (and
number of transactions though we don't need to specify this) as it
doesn't require weighting.

Now do the same for service-kpi