Split backdrop url between internal and external
The current backdrop url is backed by fastly, which means server side
requests are going out onto the internet to contact an app which is
running within the same infrastructure. To avoid this we introduce
config for an externalUrl, and when running client side use this url,
but when running server side, use the internal url.