Add performance timing to data and rendering
We need to measure and understand where Spotlight is spending most of
it's time when rendering a page. This introduces some functions to
measure the performance for plain code and event triggering. It only
runs on the server as this is where we are interested in tracking
performance at the moment.