wip
I don't think this is a priority

we need to sort the url pushed by date picker to be real query params
we need to have the server recognise and use these params to make the
json url
we need datepicker to rerender this url and also the table - a big ask
for components that may only exist server side.
maybe we just need to reload the page after the url interpreting is
done?