Fix tables displayed instead of graphs when javascript is turned off
All of the graphs on the dashboards are underpinned by a table.
If javascript is turned off the table is displayed instead of the graph.

The hack added to add a department name had broken this feature. The
code that renders tables is shared across spotlight. When the department
name was added to the table rendering code, it was also added behind the
scenes to every graph. The graphs still rendered correctly, but when
javascript was turned off, the table failed to render.

Added a  fix to allow the x axis of the graph/table to accept a list or
a single value. This is because whilst some of the config for the x and
y axes is held in spotlight, the majority of it is held in the module
config in stagecraft and can only be edited via the admin app.