Cache template sources
Express will cache the result of a call to render() but prior to this call
we make a call to fs.readFileSync for the mustache or underscore template.
This means a lot of IO as the static template source is never cached.
This commit adds memory-cache so that the template cache is accessible,
the cached templates are keyed by path. No expiry time was added but this
is possible as a third arg to cache.put() if it proves necessary.