Upgrade jQuery and give it a DOM object
Whereas with jQuery 1.8.3, it was possible to just require jquery
and away you went, newer versions of jQuery need a window object
when they are started up in node.

Creating a new DOM using jsdom and passing it through to jquery
fixes most of the problems we were having.

source: https://stackoverflow.com/questions/20380958/browserify-with-jquery-2-produces-jquery-requires-a-window-with-a-document