Rewrite table code to avoid using global jQuery
This is a super weird bug. Since upgrading jQuery, we've had to
pass it a window object right when the app is starting for
jQuery to instantiate properly.

Unfortunately, in our _tests_ it seems like we have a global
version of jQuery floating around which is different to the
global version of jQuery assumed by the code in `table.js`.

Whereas the running app itself has no problem with the jQuerys
in `table.js`, running the tests causes some error whereby the
local and global versions of jQuery being used by table.js
(`$()` vs `this.$()`) have different DOMs from each other.

This causes the tests to fail even though the app itself is fine.

Here I've just rewritten the code to always use the table view's
jQuery (`this.$()`).

Behaviour is the same as before, and now the tests are happy.