No noscript tags
So it looks like, as is common with Backbone, we have really
skimpy template files which Backbone will then populate once we
have actual data to fill them with. Easy.

Unfortunately, after upgrading jQuery and getting the tests to
pass, I ran into a problem where the fallback tables were missing
(The fallback tables are (sortable) data tables you see instead
of the interactive graphs if javascript is turned off.) They were
hidden behind `<noscript>` tags so that they would only show
up in browsers where scripting was missing/disabled.

Somehow, upgrading jQuery caused all the fallback tables to
disappear, which took me forever to diagnose.

Here's the deal: at some point between jQuery versions, jQuery
became unable to select things inside of `<noscript>` tags.
This generally makes sense:

- if scripting is permitted, anything wrapped in a `<noscript>`
  will not appear in the browser, so (in most cases) it makes
  sense for jQuery to ignore it
- if scripting is disabled, jQuery will not run anyway, so it can't
  do any DOM manipulation full stop

In our case however, we are using Backbone both on the server
and on the client, which is pretty unusual. This means that when
the pages are being built on the server, jQuery is not able to
find the `.visualisation-table-fallback` element because it is
wrapped with a `<noscript>` tag.

Endlessly frustrating, tracking down this bug.

Anyway, the solution I went for was just to remove all instances
of `<noscript>` tags and instead add a CSS rule that hides all
visualisation tables when the `.js-enabled` class has been set on
the `<body>` element. This should have exactly the same intended
effect.

- when scripting is enabled, the fallback tables will be hidden
  and replaced by interactive tables
- when scripting is disabled, the class will not be set on the
  `<body>` element and the fallback tables will be visible

Simple fix but it took some time to track down.