Initiate JSONSummary view once collection exists
On other pages, we are waiting until the synced data has
been retrieved before trying to initialize views that presume that
data.

For some reason that is still a bit obscure to me, the tab view
is rendering itself and its subviews before syncing the data from
the server, which then breaks when the JSONSummary assumes the
data exists.

This change will only init the JSONSummary View once the collection
exists, which avoids the error we were getting, but it also means
the tabbed interface is still partially building itself on pages
where JS has been disabled.

If we want to remove it altogether, that's another story.