Update vendor/helper libraries used for testing
Updated 1 library and 2 helpers

- app/vendor/jquery
    - from: http://code.jquery.com/jquery-1.11.1.js
- spec/helpers/jasmine-jquery_helper
    - from: https://www.npmjs.com/package/jasmine-jquery
- spec/helpers/moment_helper

The first three I just dropped in based on googling for the most
current version of the file. The fact that they're not package
managed is going to cause a similar issue in 3 years when they're
once again super out of date.

The `moment_helper.js` one wasn't a vendored file, but one that's
been written just for use in this project. Accordingly,  I've just
rewritten it based on the instructions at:
https://jasmine.github.io/2.0/upgrading.html