Prefix helper files for client/server tests
So this is the hilarious result of isomorphic javascript.

Here's the deal:

`grunt jasmine` tests were using jasmine 1.x and phantom <2
`grunt jasmine_node` tests *are* using jasmine 1.x but not phantom

Since the phantomjs (`jasmine`) tests weren't running on node 6,
I had to upgrade the library to get phantom to run at all.
This meant that we were then using Jasmine 2.x.
Accordingly, I updated all of the tests in spec/client to the new
syntax and updated some of the dependencies and the helper files.

At the same time, because the `jasmine_node` tests continued to
run successfully, I never changed them. Don't fix what ain't broke.

So now the `jasmine_node` tests are all still working, but they're
running jasmine 1.x, which means they need some of the old helper
files.
By contrast, the `jasmine` tests are all newly working, but
they're running jasmine 2.x and they need the updated helper files.

This means I've prefixed all the helper setup files with either
`client-`, `server-`, or `shared-` and updated the jasmine
configuration in the Gruntfile for both sets of tests to only
look for the right set of helper files.

//

Another change is that, before, the tests in the `/spec/shared/`
folder would be run by both the `jasmine` and the `jasmine_node`
tests. Since they are all written for Jasmine 1.x, they no
longer run in `jasmine` test run, and I'm just going to leave it
that way.