window.history.replaceState was unspiable
Whether it's a new jasmine thing or whatever, it seems like it's
no longer possible to overwrite a funtion on the native `window`
object.

Source: https://stackoverflow.com/questions/8919370/jasmine-mock-window-object

So, as a (probably not ideal, but I've kind of had it with these
tests) workaround, we can just check the window.location.search
property directly and see that it is equal to the string we
expect.

In order that the `.location` doesn't become cluttered with other
parameters, we're truncating query parameters in the beforeEach
function.
This was also a problem for us in another file, but truncating
the query parameters there works equally well.