Streamline `jasmine` specs in gruntfile
Since we are including every file except for one, it's cleaner and
less ambigious to just use a global file matching pattern instead
of listing out the various directories.

Also, I'm stopping the SpecRunner.html from being saved by default
because it was useful when debugging but probably not actually
needed outside of that context.