Fix link updating behaviour
The previous code was:
1. removing the <li> element and replacing it with <a> elements
2. adding more <a> elements without removing the past ones

This meant that on some pages, where it was possible to filter
data in the visualisation, modifying the filters would result
in a bunch of links being appended one after another without
cleaning up the previous links.
(for example)[https://www.gov.uk/performance/check-driving-information/volumetrics]

This new behaviour works as intended, the code is simpler, and
the tests are useful.