Remove CSV link on individual module pages
Some of our CSVs are having formatting issues, which leaves us with
several options:

1. fix the formatting issue through backdrop
2. fix the formatting issue by returning CSVs in spotlight
3. remove the CSV links
4. do nothing -- leave everything as it is

We decided that (1) doesn't really solve the problem and that (2)
would be too hard for something we don't really know if people
care about at all.

Going with (3) until we have a compelling reason not to.

//

Note that it is still possible to get the CSV response -- all we
are doing is hiding the link. To get the CSV response instead
of the JSON one:
- copy the JSON link
- change where it says "format=json" to "format=csv"
- ta-da! same as before