Add documentation about how spotlight handles requests
The genesis of this is that I've spent a bunch of time tracing
through what to me is a labyrinthine application structure.

I can intellectually appreciate that spotlight has a cool,
avant-garde application design or something but, on a practical
level, the thing is hella complicated and all the super smart
people who designed it have left.

Hopefully this is more helpful than the very high level overview
that currently exists in the /docs folder.