Remove cheapseats from integration tests
We've had continual problems with cheapseats failing randomly.
And then we've also had a genre of problems where cheapseats
fails non-randomly, but almost every time it has been because
cheapseats was misconfigured or cheapseats needed to be updated.

It's rare that cheapseats actually finds legitimate problems
for us. It's not that it has never found anything -- it's more
that the work done keeping cheapseats running feels
disproportionate to the amount of bugs it uncovers.