Remove redundant packages
Both of these packages are themselves dependencies of grunt tasks.
That is,
- `node-sass` is a dependency of `grunt-sass`
- `jshint` is a dependency of `grunt-contrib-jshint`

Under npm < 4, the way of organising dependencies in the
`/node_modules` folder was different than now. I assume that for
some reason they needed a predictable path to these dependencies,
which we have now that we're using a more up-to-date version of npm.

This means that we can remove these downstream dependencies, as
they are already going to be brought in by our grunt packages.