Switch to using the node_buildpack rather than docker
Rather than create, push, pull, push a docker image this PR instead uses
the node_buildpack for PaaS.  It also introduces zero-downtime deploy so
that we do not have downtime when the previous image is destroyed.

We have added one manifest per space, partly to be explicit and allow
scaling at different levels, but also partly because the BGD plugin
wants to create a route with the app's name, even if we specify one in
the manifest.  This would cause problems when pushing to a different
space because the hostname would be in use.