make manifests match reality
The manifests for spotlight set one name for the app, but
etc/deploy.sh overrides this to a different name.

This is confusing and has left us with multiple copies of spotlight
running, bound to the same route.

To avoid this footgun in future, I think we should have the manifests
use the actual app name that we actually deploy.  I hope this is not a
controversial opinion.