adds download form to export historical data
displays a small HTML form to select a start/end month/year and file
format, that when submitted requests an export of the data for the
current page for the given time range.

without the graph functionality, it is not easy to fetch the historical
data that is available for the various service dashboards, having a form
makes this more discoverable.

The addition of the CSV format, lowers the bar to who might be able to
consume the data, since it is more readily supported by common
spreadsheet tools.